backend/requirements.txt
fastapi
uvicorn[standard]
biopython
requests
numpy
python-multipart
fastapi-cors 